feat: setup repo for futher work. Add linter, prettier, husky
