+++ .editorconfig
@@ -1,0 +1,16 @@
+root = true
+
+[*]
+end_of_line = lf
+insert_final_newline = true
+trim_trailing_whitespace = true
+
+[*.{ts, tsx}]
+charset = utf-8
+indent_style = tab
+indent_size = 2
+rulers = 80, 100, 120
+
+[{package.json}]
+indent_style = space
+indent_size = 2
