+++ .prettierrc.json
@@ -1,0 +1,10 @@
+{
+  "endOfLine": "auto",
+  "semi": true,
+  "singleQuote": true,
+  "tabWidth": 2,
+  "trailingComma": "es5",
+  "printWidth": 100,
+  "arrowParens": "always",
+  "plugins": ["prettier-plugin-tailwindcss"]
+}
