+++ app/layout.tsx
@@ -1,0 +1,23 @@
+import { Inter } from 'next/font/google';
+
+import type { Metadata } from 'next';
+import './globals.css';
+
+const inter = Inter({ subsets: ['latin'] });
+
+export const metadata: Metadata = {
+  title: 'Alema consulting',
+  description: 'Website related to consulting service Alema, located in Berlin',
+};
+
+export default function RootLayout({
+  children,
+}: Readonly<{
+  children: React.ReactNode;
+}>) {
+  return (
+    <html lang="en">
+      <body className={inter.className}>{children}</body>
+    </html>
+  );
+}
