+++ app/page.tsx
@@ -1,0 +1,7 @@
+export default function Home() {
+  return (
+    <main className="">
+      <div className="">Hello Alema!</div>
+    </main>
+  );
+}
