+++ package.json
@@ -1,0 +1,52 @@
+{
+  "name": "alema_consulting",
+  "version": "0.1.0",
+  "private": true,
+  "scripts": {
+    "dev": "next dev",
+    "build": "next build",
+    "start": "next start",
+    "type-check": "tsc --noEmit",
+    "eslint": "eslint --ext .js --ext .jsx --ext .ts --ext .tsx ./",
+    "eslint:fix": "eslint --fix --ext .js --ext .jsx --ext .ts --ext .tsx ./",
+    "format:fix": "prettier --write \"./**/*.{ts,tsx,css}\"",
+    "prepare": "husky install"
+  },
+  "dependencies": {
+    "gray-matter": "^4.0.3",
+    "markdown-to-jsx": "^7.4.1",
+    "next": "14.1.0",
+    "react": "^18",
+    "react-dom": "^18"
+  },
+  "devDependencies": {
+    "@types/node": "^20",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
+    "@typescript-eslint/eslint-plugin": "^7.0.1",
+    "@typescript-eslint/parser": "^7.0.1",
+    "autoprefixer": "^10.0.1",
+    "eslint": "^8.56.0",
+    "eslint-config-airbnb": "^19.0.4",
+    "eslint-config-next": "14.1.0",
+    "eslint-config-prettier": "^9.1.0",
+    "eslint-import-resolver-typescript": "^3.6.1",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsx-a11y": "^6.8.0",
+    "eslint-plugin-prettier": "^5.1.3",
+    "eslint-plugin-react": "^7.33.2",
+    "eslint-plugin-react-hooks": "^4.6.0",
+    "eslint-plugin-simple-import-sort": "^12.0.0",
+    "husky": "^8.0.0",
+    "lint-staged": "^15.2.2",
+    "postcss": "^8",
+    "prettier": "^3.2.5",
+    "prettier-plugin-tailwindcss": "^0.5.11",
+    "tailwindcss": "^3.3.0",
+    "typescript": "^5"
+  },
+  "lint-staged": {
+    "./src/**/*.{tsx, ts}": "eslint",
+    "./src/**/*.{tsx, ts, html, css}": "prettier --write"
+  }
+}
